完善映射关系删除功能

+++ src/com/woting/crawlerdb/dict/service/CDictService.java
@@ -10,14 +10,19 @@ import java.util.Map;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.Resource;
+import javax.servlet.ServletContext;
 
+import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
 import org.springframework.stereotype.Service;
+import org.springframework.web.context.support.WebApplicationContextUtils;
 
+import com.spiritdata.framework.FConstants;
 import com.spiritdata.framework.core.cache.CacheEle;
 import com.spiritdata.framework.core.cache.SystemCache;
 import com.spiritdata.framework.core.dao.mybatis.MybatisDAO;
 import com.spiritdata.framework.core.model.tree.TreeNode;
 import com.spiritdata.framework.core.model.tree.TreeNodeBean;
+import com.spiritdata.framework.ext.spring.redis.RedisOperService;
 import com.spiritdata.framework.util.JsonUtils;
 import com.spiritdata.framework.util.SequenceUUID;
 import com.spiritdata.framework.util.TreeUtils;
@@ -375,15 +380,41 @@ public class CDictService {
 	 * @param id
 	 * @return
 	 */
-	public boolean delDictResRef(String ids, boolean isOrNoRemove) {
+	public List<Map<String, Object>> delDictResRef(String ids, boolean isOrNoRemove) {
 		String[] chamapids = ids.split(",");
 		if (chamapids!=null && chamapids.length>0) {
+			List<Map<String, Object>> retLs = new ArrayList<>();
+			RedisOperService redis = null;
+			ServletContext sc=(SystemCache.getCache(FConstants.SERVLET_CONTEXT)==null?null:(ServletContext)SystemCache.getCache(FConstants.SERVLET_CONTEXT).getContent());
+	        if (WebApplicationContextUtils.getWebApplicationContext(sc)!=null) {
+	            JedisConnectionFactory js =(JedisConnectionFactory) WebApplicationContextUtils.getWebApplicationContext(sc).getBean("connectionFactory123");
+	            redis = new RedisOperService(js, 6);
+	        }
 			for (String id : chamapids) {
-				crawlerService.deleteCCateResRef(id);
+				if (redis.get("wt_ChannelMap_Ref_"+id)!=null) {
+					Map<String, Object> m = new HashMap<>();
+					m.put("PerId", id);
+					m.put("Message", "删除失败，关系正在处理");
+					retLs.add(m);
+				} else {
+					if (isOrNoRemove) { // 删除栏目关系表数据
+						crawlerService.deleteCCateResRef(id);
+						Map<String, Object> m = new HashMap<>();
+						m.put("PerId", id);
+						m.put("Message", "开始关系执行关系删除");
+						retLs.add(m);
+					} else { // 只删除对应关系
+						channelMapService.deleteById(id);
+						Map<String, Object> m = new HashMap<>();
+						m.put("PerId", id);
+						m.put("Message", "关系删除成功");
+						retLs.add(m);
+					}
+				}
 			}
-			return true;
+			return retLs;
 		}
-		return false;
+		return null;
 	}
 	
 	public boolean saveCrawlerFile() {
